refactor(topic metrics): replace legacy topic metrics UI
- promote the collection-based UI to the existing topic metrics route
- preserve the topic list shortcut and prefill topic filters on creation
- remove the duplicate collections entry and obsolete wildcard restrictions
