+++ src/common/featureGate.ts
@@ -18,11 +18,7 @@ export const FEATURE_ROUTE_MAP: Record<FeatureName, string[]> = {
   [FeatureName.Gateways]: ['/gateway'],
   [FeatureName.GCPDevice]: [],
   [FeatureName.MessageTransformation]: ['/message-transform'],
-  [FeatureName.Metrics]: [
-    '/topic-metrics',
-    '/topic-metrics-collections',
-    '/monitoring/integration',
-  ],
+  [FeatureName.Metrics]: ['/topic-metrics', '/monitoring/integration'],
   [FeatureName.MQTTExtensions]: [
     '/delayed-pub',
     '/delayed-pub-configuration',
