+++ src/router/index.ts
@@ -1009,22 +1009,6 @@ export const routes: Array<RouteRecordRaw> = [
       },
     ],
   },
-  // Topic Metrics Collections
-  {
-    path: '/topic-metrics-collections',
-    component: Layout,
-    meta: {
-      hideKey: 'topicMetricsCollections',
-      authRequired: true,
-    },
-    children: [
-      {
-        path: '',
-        name: 'topic-metrics-collections',
-        component: () => import('@/views/Diagnose/TopicMetricsCollections.vue'),
-      },
-    ],
-  },
   // Slow Sub
   {
     path: '/slow-sub',
